- implement rating - prevent rating, tagging, etc. of virtual directories - make jumping to songs from playlist in filelist work also for songs   in compilations - some cleanups

+++ src/services/songdbs/sqlite.py
@@ -94,7 +94,7 @@ CREATE TABLE songs (
   date_updated          TIMESTAMP,
   date_lastplayed       TIMESTAMP,
   playcount             INTEGER,
-  rating                FLOAT
+  rating                INTEGER
 );
 
 CREATE INDEX album_id ON albums(name);
